shallow water equation solver (#68)
* shallow water equation solver

* differential operators from voronoi diagram

* new facet implementation

* more work on swe

* shallow water test cases

* fix compile error

* include filesystem

* use original operators method

* commit to new facet implementation

* fix some warnings

* make nlopt optional, fix some more warnings

* forgot to wrap around nlopt include

* more warning fixes

* more warning fixes

* make abseil an option

* ignore coverage mismatch

* revert coverage option

* try to ignore-errors mismatch

* ignore coverage for now

* forgot to skip coverage

+++ src/device_util.h
@@ -211,7 +211,7 @@ class device_hash_set {
     new_data = (T*)malloc(sizeof(T) * capacity);
     new_taken = (bool*)malloc(sizeof(bool) * capacity);
 #endif
-    for (int k = 0; k < capacity; k++) {
+    for (size_t k = 0; k < capacity; k++) {
       new_taken[k] = false;
       new_data[k] = std::numeric_limits<T>::max();
     }
