+++ src/neighbors.h
@@ -61,7 +61,7 @@ class queue {
 };
 
 struct NearestNeighborsWorkspace {
-  NearestNeighborsWorkspace(int k, int nmax = 200) : n_neighbors(k) {
+  NearestNeighborsWorkspace(size_t k, size_t nmax = 200) : n_neighbors(k) {
     neighbors.reserve(nmax);
   }
 
@@ -96,7 +96,7 @@ struct NearestNeighborsWorkspace {
     neighbors.insert({n, l});
   }
 
-  const int n_neighbors;
+  const size_t n_neighbors;
   std::unordered_map<uint32_t, uint8_t> neighbors;
   queue<std::pair<uint32_t, double>> sites;
   size_t total_neighbors{0};
@@ -145,7 +145,7 @@ struct SphereQuadtreeWorkspace {
   void add(uint32_t n, double d) { neighbors.emplace_back(n, d); }
 
   std::vector<std::pair<uint32_t, double>> neighbors;
-  const int n_neighbors{0};
+  const size_t n_neighbors{0};
 };
 
 class SphereQuadtree {
